Framework version is read from POM
Framework version for stats purposes was defined in constant. The
constant had to be manually changed as part of every release. This is
major issue for continuous integration/deployment process.

Currently the version is read from property file, where is updated
automatically by maven resource module. Allowing resource filtering in
pom.xml enables variable substitution.

Class VersionReader (and tests for its method) has been created to allow
reading from the file. For now the read results are not cached, as they
are used in a single place that do caching itself.